requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
flask-mysql-connector==1.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
wsproto==1.2.0
